refactoring sequence mocking, adding more mock sequences
