+++ pipeline/common/test_variant_utils.py
@@ -1,7 +1,7 @@
 
 import pytest
-from common import seq_utils, variant_utils
-from common.hgvs_utils import HgvsWrapper
+
+from common import variant_utils
 
 variants = ['chr17:g.43094426:C>T',
             'chr17:g.43071235:CCCT>C',
@@ -10,25 +10,14 @@ variants = ['chr17:g.43094426:C>T',
             'chr17:g.43125412:GCT>GAG']
 
 
-@pytest.fixture(scope="module")
-def wrapper():
-    return HgvsWrapper()
-
-
-@pytest.fixture(scope="module")
-def seq_fetcher():
-    return seq_utils.SeqRepoWrapper.get_instance()
-
-
 @pytest.mark.parametrize("v", variants)
 def test_variant_simple_roundtripping(v):
      assert str(variant_utils.VCFVariant.from_str(v)) == v
 
 
 @pytest.mark.parametrize("v", variants)
-def test_variant_hgvs_roundtripping(v):
-    hgvs_wrap = HgvsWrapper()
-    hgvs = variant_utils.VCFVariant.from_str(v).to_hgvs_obj(hgvs_wrap.contig_maps[hgvs_wrap.GRCh38_Assem])
+def test_variant_hgvs_roundtripping(v, hgvs_wrapper):
+    hgvs = variant_utils.VCFVariant.from_str(v).to_hgvs_obj(hgvs_wrapper.contig_maps[hgvs_wrapper.GRCh38_Assem])
     assert str(variant_utils.VCFVariant.from_hgvs_obj(hgvs)) == v
 
 
@@ -39,8 +28,8 @@ def test_variant_hgvs_roundtripping(v):
                           ('NC_000017.11:g.43045730_43045731insCAGTGCCAGGACAGCAGG', 'chr17:g.43045730:C>CCAGTGCCAGGACAGCAGG'),
                           ('NC_000017.11:g.43091784_43091817del', 'chr17:g.43091783:CGGTAGCAACGGTGCTATGCCTAGTAGACTGAGAA>C')
                         ])
-def test_hgvs_to_vcf(wrapper, v, expected, seq_fetcher):
-    hgvs_var = wrapper.hgvs_parser.parse(v)
+def test_hgvs_to_vcf(hgvs_wrapper, v, expected, seq_fetcher):
+    hgvs_var = hgvs_wrapper.hgvs_parser.parse(v)
     v = variant_utils.VCFVariant.from_hgvs_obj(hgvs_var, seq_fetcher)
 
     assert str(v) == expected
