+++ pipeline/data/mock_fetch_seq-NC_000013.11-32340944-32340945.txt
@@ -1,0 +1,1 @@
+C
