+++ pipeline/data/mock_fetch_seq-NC_000017.11-43045729-43045730.txt
@@ -1,0 +1,1 @@
+C
