+++ pipeline/data/mock_fetch_seq-NC_000017.11-43086108-43086109.txt
@@ -1,0 +1,1 @@
+T
