+++ pipeline/data/mock_fetch_seq-NC_000017.11-43091782-43091783.txt
@@ -1,0 +1,1 @@
+C
