+++ pipeline/data/mock_fetch_seq-NC_000017.11-43091783-43091817.txt
@@ -1,0 +1,1 @@
+GGTAGCAACGGTGCTATGCCTAGTAGACTGAGAA
