+++ pipeline/data/mock_fetch_seq-NC_000017.11-43098660-43098661.txt
@@ -1,0 +1,1 @@
+C
