+++ pipeline/data_merging/test_variant_merging.py
@@ -1,17 +1,18 @@
 import itertools
-import glob
 import os
 import unittest
 
-import bioutils.seqfetcher
-from mock import patch
 import pytest
+
+import bioutils
 import vcf
 from hypothesis import given, assume, settings
 from hypothesis.strategies import integers, tuples, sampled_from, lists
+from mock import patch
 
 from common import seq_utils
 from common.config import load_config, extract_gene_regions_dict
+from common.conftest import seq_fetcher, fetch_seq_mock_data
 from utilities import round_sigfigs
 from variant_equivalence import variant_equal, find_equivalent_variant, find_equivalent_variants_whole_seq
 from variant_merging import normalize_values, add_variant_to_dict, \
@@ -389,50 +390,35 @@ class TestVariantMerging(unittest.TestCase):
                 except ValueError:
                     self.assertEqual(val, '-')
 
-@pytest.fixture()
-def fetch_seq_mock_data():
-    mock_data = {}
-
-    # mock data was generated by running tests using 'make test' and
-    # saving the output of SeqRepoWrapper._fetch_seq into the corresponding files
-    for path in glob.glob(os.path.join(data_dir, 'mock_fetch_seq*')):
-        with open(path, 'r') as f:
-            fn = os.path.basename(path).rstrip('.txt')
-            key = tuple(fn.split('-')[-3:]) # ac, start, end
-            mock_data[key] = f.readline().strip()
-
-    return mock_data
-
-
-def test_find_equivalent_variant(fetch_seq_mock_data):
+def test_find_equivalent_variant(seq_fetcher):
     # mocking _fetch_seq method
-    with patch.object(bioutils.seqfetcher, 'fetch_seq', side_effect=lambda ac, s, e: fetch_seq_mock_data[(str(ac), str(s), str(e))]):
-        seq_wrapper = seq_utils.SeqRepoWrapper()
-        # empty case
-        assert [] == find_equivalent_variant({}, seq_wrapper)
+    #with patch.object(bioutils.seqfetcher, 'fetch_seq', side_effect=lambda ac, s, e: fetch_seq_mock_data[(str(ac), str(s), str(e))]):
+    seq_wrapper = seq_fetcher
+    # empty case
+    assert [] == find_equivalent_variant({}, seq_wrapper)
 
-        # a bunch of variants. If they appear in the same set, they are considered equivalent
-        example_variants = [
-            frozenset({'chr13:g.32355030:A>AA'}),
-            frozenset({'chr13:g.32339774:GAT>G', 'chr13:g.32339776:TAT>T'}),
-            frozenset({'chr17:g.43090921:G>GCA', 'chr17:g.43090921:GCA>GCACA'})
-        ]
+    # a bunch of variants. If they appear in the same set, they are considered equivalent
+    example_variants = [
+        frozenset({'chr13:g.32355030:A>AA'}),
+        frozenset({'chr13:g.32339774:GAT>G', 'chr13:g.32339776:TAT>T'}),
+        frozenset({'chr17:g.43090921:G>GCA', 'chr17:g.43090921:GCA>GCACA'})
+    ]
 
-        # construct variant dict (flattening example_variants!)
-        variant_dict = {v: VCFVariant(
-            int(v.split(':')[0].lstrip('chr')),
-            int(v.split(':')[1].lstrip('g.')),
-            v.split(':')[2].split('>')[0],
-            v.split(':')[2].split('>')[1]) for eq_variants in example_variants for v
-        in
-            eq_variants
-        }
+    # construct variant dict (flattening example_variants!)
+    variant_dict = {v: VCFVariant(
+        int(v.split(':')[0].lstrip('chr')),
+        int(v.split(':')[1].lstrip('g.')),
+        v.split(':')[2].split('>')[0],
+        v.split(':')[2].split('>')[1]) for eq_variants in example_variants for v
+    in
+        eq_variants
+    }
 
-        margin = 20
-        chunk_provider = seq_utils.ChunkBasedSeqProvider(variant_dict.values(), margin, seq_wrapper)
+    margin = 20
+    chunk_provider = seq_utils.ChunkBasedSeqProvider(variant_dict.values(), margin, seq_wrapper)
 
-        assert frozenset(example_variants) == frozenset(
-            find_equivalent_variant(variant_dict, chunk_provider))
+    assert frozenset(example_variants) == frozenset(
+        find_equivalent_variant(variant_dict, chunk_provider))
 
 
 def test_find_equivalent_variant_whole_seq(fetch_seq_mock_data):
